requirements.txt
tensorflow>=2.10
numpy
pandas
opencv-python
matplotlib
scikit-learn
keras
google-colab; platform_system == 'Linux'
